requirements.txt
tensorflow
tensorflow-hub
imagehash
numpy
scipy
Pillow